requirements.txt
azure-common==1.1.28
azure-core==1.23.1
azure-core-tracing-opentelemetry==1.0.0b9
azure-identity==1.9.0
azure-mgmt-compute==26.1.0
azure-mgmt-core==1.3.0
backoff==1.11.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.0
Deprecated==1.2.13
googleapis-common-protos==1.56.0
grpcio==1.44.0
idna==3.3
isodate==0.6.1
msal==1.17.0
msal-extensions==0.3.1
msrest==0.6.21
oauthlib==3.2.0
opentelemetry-api==1.11.1
opentelemetry-exporter-otlp-proto-grpc==1.11.1
opentelemetry-proto==1.11.1
opentelemetry-sdk==1.11.1
opentelemetry-semantic-conventions==0.30b1
portalocker==2.4.0
protobuf==3.20.1
pycparser==2.21
PyJWT==2.3.0
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
typing_extensions==4.2.0
urllib3==1.26.9
wrapt==1.14.0
